Use (str).split more correctly
The second argument to split is the maximum number of splits, not
the maximum number of output segments.

Because the URLs (and hence the slug) of blog post files are
constrained by the URLs, and in any case because we used an overly-
loose exception handler, the resultant ValueError from trying to
unpack too many values into the lvar tuples would have been
handled anyway. We now catch just ValueError, so that would still
be swallowed if for some reason you constructed your own URL
scheme.